+++ app/src/main/java/hous/release/android/presentation/withdraw/feedback/FeedbackFragment.kt
@@ -1,4 +1,4 @@
-package hous.release.android.presentation.withdraw
+package hous.release.android.presentation.withdraw.feedback
 
 import android.annotation.SuppressLint
 import android.os.Bundle
