+++ app/src/main/res/layout/fragment_feedback.xml
@@ -7,14 +7,14 @@
 
         <variable
             name="viewModel"
-            type="hous.release.android.presentation.withdraw.FeedbackViewModel" />
+            type="hous.release.android.presentation.withdraw.feedback.FeedbackViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/cl_feedback"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".presentation.withdraw.FeedbackFragment">
+        tools:context=".presentation.withdraw.feedback.FeedbackFragment">
 
         <ImageButton
             android:id="@+id/btn_feedback_back"
