+++ app/src/main/res/navigation/nav_withdraw.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/feedbackFragment"
-        android:name="hous.release.android.presentation.withdraw.FeedbackFragment"
+        android:name="hous.release.android.presentation.withdraw.feedback.FeedbackFragment"
         android:label="FeedbackFragment"
         tools:layout="@layout/fragment_feedback">
         <action
@@ -17,7 +17,7 @@
 
     <fragment
         android:id="@+id/withdrawFragment"
-        android:name="hous.release.android.presentation.withdraw.WithdrawFragment"
+        android:name="hous.release.android.presentation.withdraw.withdraw.WithdrawFragment"
         android:label="WithdrawFragment"
         app:popUpTo="@id/feedbackFragment"
         app:popUpToInclusive="true"
@@ -29,7 +29,7 @@
 
     <fragment
         android:id="@+id/withdrawDoneFragment"
-        android:name="hous.release.android.presentation.withdraw.WithdrawDoneFragment"
+        android:name="hous.release.android.presentation.withdraw.withdrawDone.WithdrawDoneFragment"
         android:label="WithdrawDoneFragment"
         tools:layout="@layout/fragment_withdraw_done" />
